INTYGFV-15083: Refactor by extracting InputExpectedValuePair to a common type with generics for input and expectedValue.
