+++ support/src/testFixtures/java/se/inera/intyg/common/support/facade/testsetup/model/value/InputExpectedValuePair.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package se.inera.intyg.common.support.facade.testsetup.model.value;
+
+public class InputExpectedValuePair<T, S> {
+
+    private final T input;
+    private final S expectedValue;
+
+    public InputExpectedValuePair(T input, S expectedValue) {
+        this.input = input;
+        this.expectedValue = expectedValue;
+    }
+
+    public T getInput() {
+        return input;
+    }
+
+    public S getExpectedValue() {
+        return expectedValue;
+    }
+}
